initial commit with start of new BoundaryHandler class
