+++ Source/FieldSolver/Make.package
@@ -5,6 +5,7 @@ CEXE_sources += WarpX_QED_Field_Pushers.cpp
 ifeq ($(USE_FFT),TRUE)
   include $(WARPX_HOME)/Source/FieldSolver/SpectralSolver/Make.package
 endif
+include $(WARPX_HOME)/Source/FieldSolver/ElectrostaticSolvers/Make.package
 include $(WARPX_HOME)/Source/FieldSolver/FiniteDifferenceSolver/Make.package
 include $(WARPX_HOME)/Source/FieldSolver/MagnetostaticSolver/Make.package
 include $(WARPX_HOME)/Source/FieldSolver/ImplicitSolvers/Make.package
