+++ Source/WarpX.cpp
@@ -796,37 +796,23 @@ WarpX::ReadParameters ()
         );
 #endif
 
-        // Parse the input file for domain boundary potentials
-        const ParmParse pp_boundary("boundary");
-        bool potential_specified = false;
-        // When reading the potential at the boundary from the input file, set this flag to true if any of the potential is specified
-        potential_specified |= pp_boundary.query("potential_lo_x", m_poisson_boundary_handler.potential_xlo_str);
-        potential_specified |= pp_boundary.query("potential_hi_x", m_poisson_boundary_handler.potential_xhi_str);
-        potential_specified |= pp_boundary.query("potential_lo_y", m_poisson_boundary_handler.potential_ylo_str);
-        potential_specified |= pp_boundary.query("potential_hi_y", m_poisson_boundary_handler.potential_yhi_str);
-        potential_specified |= pp_boundary.query("potential_lo_z", m_poisson_boundary_handler.potential_zlo_str);
-        potential_specified |= pp_boundary.query("potential_hi_z", m_poisson_boundary_handler.potential_zhi_str);
-        if (m_eb_enabled) {
-            potential_specified |= pp_warpx.query("eb_potential(x,y,z,t)", m_poisson_boundary_handler.potential_eb_str);
-        }
-        m_boundary_potential_specified = potential_specified;
-        if (potential_specified & (WarpX::electromagnetic_solver_id == ElectromagneticSolverAlgo::HybridPIC)) {
-            ablastr::warn_manager::WMRecordWarning(
-                "Algorithms",
-                "The input script specifies the electric potential (phi) at the boundary, but \
-                also uses the hybrid PIC solver based on Ohm’s law. When using this solver, the \
-                electric potential does not have any impact on the simulation.",
-                ablastr::warn_manager::WarnPriority::low);
-        }
-        else if (potential_specified & (WarpX::electromagnetic_solver_id != ElectromagneticSolverAlgo::None)) {
-            ablastr::warn_manager::WMRecordWarning(
-                "Algorithms",
-                "The input script specifies the electric potential (phi) at the boundary so \
-                an initial Poisson solve will be performed.",
-                ablastr::warn_manager::WarnPriority::low);
-        }
+        // m_boundary_potential_specified = potential_specified;
+        // if (potential_specified & (WarpX::electromagnetic_solver_id == ElectromagneticSolverAlgo::HybridPIC)) {
+        //     ablastr::warn_manager::WMRecordWarning(
+        //         "Algorithms",
+        //         "The input script specifies the electric potential (phi) at the boundary, but \
+        //         also uses the hybrid PIC solver based on Ohm’s law. When using this solver, the \
+        //         electric potential does not have any impact on the simulation.",
+        //         ablastr::warn_manager::WarnPriority::low);
+        // }
+        // else if (potential_specified & (WarpX::electromagnetic_solver_id != ElectromagneticSolverAlgo::None)) {
+        //     ablastr::warn_manager::WMRecordWarning(
+        //         "Algorithms",
+        //         "The input script specifies the electric potential (phi) at the boundary so \
+        //         an initial Poisson solve will be performed.",
+        //         ablastr::warn_manager::WarnPriority::low);
+        // }
 
-        m_poisson_boundary_handler.buildParsers();
 #ifdef WARPX_DIM_RZ
         pp_boundary.query("verboncoeur_axis_correction", verboncoeur_axis_correction);
 #endif
